+++ src/cmd/eqn/funny.c
@@ -21,10 +21,10 @@ void funny(int n)
 	default:
 		ERROR "funny type %d in funny", n FATAL;
 	}
-	printf(".ds %d %s\n", yyval, f);
+	printf(".ds %d %s\n", (int)yyval, f);
 	eht[yyval] = EM(1.0, ps+Funnyps) - EM(Funnyht, ps);
 	ebase[yyval] = EM(Funnybase, ps);
 	dprintf(".\tS%d <- %s; h=%g b=%g\n", 
-		yyval, f, eht[yyval], ebase[yyval]);
+		(int)yyval, f, eht[yyval], ebase[yyval]);
 	lfont[yyval] = rfont[yyval] = ROM;
 }
