+++ src/cmd/eqn/matrix.c
@@ -70,7 +70,7 @@ void matrix(int oct)	/* matrix is list of columns */
 	lfont[yyval] = rfont[yyval] = 0;
 	dprintf(".\tmatrix S%d: r=%d, c=%d, h=%g, b=%g\n",
 		yyval,nrow,ncol,eht[yyval],ebase[yyval]);
-	printf(".ds %d \"", yyval);
+	printf(".ds %d \"", (int)yyval);
 	for( i=0; i<ncol; i++ )  {
 		printf("\\*(%d%s", val[i], i==ncol-1 ? "" : space);
 		sfree(val[i]);
